Simplify and merge components
