+++ src/App.tsx
@@ -1,12 +1,12 @@
 import * as React from "react";
 import {useEffect, useState} from "react";
 import './App.css';
-import {localize} from "./main";
+import {localize, translationToString} from "./main";
 import config from "./config";
 import Notification from "./components/Notification";
 import View from "./components/View";
 import {BrowserRouter, Route, Routes} from "react-router-dom";
-import {Home} from "./components/Home";
+import Home from "./components/Home";
 
 function App() {
   const [notifications, setNotifications] = useState([]);
@@ -20,6 +20,14 @@ function App() {
         <Route path="/" element={<Home/>}/>
         <Route path="/view" element={<View/>}/>
       </Routes>
+      <footer>
+        {translationToString({
+          en: <span>The source code is available on <a href={"https://github.com/l0drex/family-tree"}>Github</a>.</span>,
+          de: <span>Der Quellcode ist auf <a href={"https://github.com/l0drex/family-tree"}>Github</a> verfügbar.</span>
+        })}
+        <span>Impressum</span>
+        <span>Datenschutzerklärung</span>
+      </footer>
     </BrowserRouter>
   } catch (error) {
     if (error.name.startsWith("Warning")) {
