+++ src/components/Header.tsx
@@ -1,11 +1,16 @@
+import * as React from "react";
 import './Header.css';
 import {Link} from 'react-router-dom';
 import {translationToString} from "../main";
+import {parseFile} from "./Form";
+import {saveDataAndRedirect} from "./Form";
 
 function Header(props) {
+  let fileInput = React.createRef<HTMLInputElement>();
+
   return (
     <header>
-      <Link to="/" onClick={removeData}>
+      <Link to="/">
         <img src={process.env.PUBLIC_URL + "/logo.svg"} width="40" height="100%" alt={translationToString({
           en: "A smiling tree.",
           de: "Ein lächelnder Baum."
@@ -18,12 +23,18 @@ function Header(props) {
       <div>
         {props.children}
       </div>
+      <nav>
+        <form id="open-file">
+          <input type="file" hidden ref={fileInput} accept="application/json"
+                 onChange={() => parseFile(fileInput.current.files[0]).then(saveDataAndRedirect)}/>
+          <button className="icon-only" onClick={() => {
+            fileInput.current.click();
+          }}>📁
+          </button>
+        </form>
+      </nav>
     </header>
   );
-}
-
-function removeData() {
-  window.sessionStorage.removeItem("familyData")
 }
 
 export default Header;
