+++ src/components/View.tsx
@@ -8,7 +8,6 @@ import InfoPanel from "./InfoPanel";
 import Header from "./Header";
 import SearchField from "./SearchField";
 import * as React from "react";
-import FamilyPath from "./FamilyPath";
 import {Person} from "gedcomx-js";
 
 
@@ -138,7 +137,6 @@ class View extends Component<any, State> {
           <TreeView colorMode={this.state.colorMode} focus={focus} focusHidden={this.state.focusHidden}
                     onRefocus={this.onRefocus.bind(this)}/>
         </main>
-        <FamilyPath focus={focus}/>
       </>
     );
   }
